+++ src/commands/aerocloud/mod.rs
@@ -22,32 +22,36 @@ pub async fn run(
         return self::set_auth_token::run(args, config, token).await;
     }
 
-    let client =
-        http::build_aerocloud_client_from_config(&config, &args.http_timeout())?;
+    let api_client = http::build_aerocloud_client(&config, args)
+        .context("building api client")?;
+
+    let file_upload_client = http::build_file_upload_client(args)
+        .context("building file upload client")?;
 
     match subcommand {
         AeroCloudScope::SetAuthToken { .. } => Ok(()),
         AeroCloudScope::CurrentUser => {
-            self::current_user::run(args, &client).await
+            self::current_user::run(args, &api_client).await
         }
         AeroCloudScope::CurrentToken => {
-            self::current_token::run(args, &client).await
+            self::current_token::run(args, &api_client).await
         }
         AeroCloudScope::V6 { command } => match command {
             AeroCloudV6Command::ListProjects { status } => {
-                self::v6::list_projects::run(args, &client, *status).await
+                self::v6::list_projects::run(args, &api_client, *status).await
             }
             AeroCloudV6Command::CreateProject { name, description } => {
                 self::v6::create_project::run(
                     args,
-                    &client,
+                    &api_client,
                     name,
                     description.as_deref(),
                 )
                 .await
             }
             AeroCloudV6Command::DeleteProjects { project_ids } => {
-                self::v6::delete_projects::run(args, &client, project_ids).await
+                self::v6::delete_projects::run(args, &api_client, project_ids)
+                    .await
             }
             AeroCloudV6Command::ListSimulations {
                 project_id,
@@ -59,7 +63,7 @@ pub async fn run(
             } => {
                 self::v6::list_simulations::run(
                     args,
-                    &client,
+                    &api_client,
                     project_id,
                     *show_results,
                     *status,
@@ -70,25 +74,30 @@ pub async fn run(
                 .await
             }
             AeroCloudV6Command::DeleteSimulations { simulation_ids } => {
-                self::v6::delete_simulations::run(args, &client, simulation_ids)
-                    .await
+                self::v6::delete_simulations::run(
+                    args,
+                    &api_client,
+                    simulation_ids,
+                )
+                .await
             }
         },
         AeroCloudScope::V7 { command } => match command {
             AeroCloudV7Command::ListProjects { status } => {
-                self::v7::list_projects::run(args, &client, *status).await
+                self::v7::list_projects::run(args, &api_client, *status).await
             }
             AeroCloudV7Command::CreateProject { name, description } => {
                 self::v7::create_project::run(
                     args,
-                    &client,
+                    &api_client,
                     name,
                     description.as_deref(),
                 )
                 .await
             }
             AeroCloudV7Command::DeleteProjects { project_ids } => {
-                self::v7::delete_projects::run(args, &client, project_ids).await
+                self::v7::delete_projects::run(args, &api_client, project_ids)
+                    .await
             }
             AeroCloudV7Command::ListSimulations {
                 project_id,
@@ -100,7 +109,7 @@ pub async fn run(
             } => {
                 self::v7::list_simulations::run(
                     args,
-                    &client,
+                    &api_client,
                     project_id,
                     *show_results,
                     *status,
@@ -111,12 +120,13 @@ pub async fn run(
                 .await
             }
             AeroCloudV7Command::ListReusableModels => {
-                self::v7::list_reusable_models::run(args, &client).await
+                self::v7::list_reusable_models::run(args, &api_client).await
             }
             AeroCloudV7Command::CreateModel { params } => {
                 self::v7::create_model::run(
                     args,
-                    &client,
+                    &api_client,
+                    &file_upload_client,
                     &params
                         .clone()
                         .contents()
@@ -131,7 +141,7 @@ pub async fn run(
             } => {
                 self::v7::create_simulation::run(
                     args,
-                    &client,
+                    &api_client,
                     model_id.clone(),
                     project_id.clone(),
                     &params
@@ -142,11 +152,15 @@ pub async fn run(
                 .await
             }
             AeroCloudV7Command::DeleteSimulations { simulation_ids } => {
-                self::v7::delete_simulations::run(args, &client, simulation_ids)
-                    .await
+                self::v7::delete_simulations::run(
+                    args,
+                    &api_client,
+                    simulation_ids,
+                )
+                .await
             }
             AeroCloudV7Command::WaitForSimulations { ids } => {
-                self::v7::wait_for_simulations::run(args, &client, ids).await
+                self::v7::wait_for_simulations::run(args, &api_client, ids).await
             }
             AeroCloudV7Command::Batch { root_dir } => {
                 if args.debug && args.log_to_path.is_none() {
@@ -156,7 +170,8 @@ pub async fn run(
                 }
 
                 self::v7::batch::run(
-                    &client,
+                    &api_client,
+                    &file_upload_client,
                     root_dir.as_ref().map(PathBuf::as_path),
                 )
                 .await